fix: update JS package
