+++ packages/js/src/constants.ts
@@ -4,6 +4,4 @@ export const TRACK_TYPES = {
   TRACK_LINK: 'trackLink',
   TRACK_SEARCH: 'trackSiteSearch',
   TRACK_VIEW: 'trackPageView',
-  TRACK_ECOMMERCE_ORDER: 'trackEcommerceOrder',
-  TRACK_ECOMMERCE_CART_UPDATE: 'trackEcommerceCartUpdate',
 }
